+++ NEWS.md
@@ -2,6 +2,14 @@
 
 Version numbers follow [Semantic Versioning](https://semver.org/).
 
+# portalcasting 0.32.0
+*In Progress*
+
+### Relocation of prefab controls 
+* Moved from source code scripts to `.yaml` files in `inst/extdata`
+
+
+
 # [portalcasting 0.31.0](https://github.com/weecology/portalcasting/releases/tag/v0.31.0)
 *2022-04-01*